server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.6.3",
    "fastapi>=0.115.13",
    "groq>=0.28.0",
    "uvicorn>=0.34.3",
    "python-dotenv>=1.0.0",
]
